Added basic sanity check for all builds and tests (intel only)
Added a script that can be used to do a basic verify that all platforms and
tests build/run successfully. Not a guarantee of no bugs but should pick up
those 'oh I forgot to run the tests on watchOS before commit' errors which
happen more than I'd like to admit.