Removed unused localizations
Cocoapods had an issue where the localizations defined in the package were overriding
the localizations in the project that QRCode was being imported into (the default 'resources'
behaviour for cocoapods seems to insert the resources at the root level of the project).

Changing to use resource_bundles instead I had trouble getting working (this fixed the
apps localizations, but then QRCode localizations didn't work. Very odd!)

Given that there's only a single localization in QRCode at the moment, I'm removing
this functionality.

Supporting Cocoapods and SPM at the same time in the same package is _hard_.